+++ cmd/backendcmd/spawn_server.go
@@ -1,14 +1,17 @@
 // Copyright (C) 2022, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
-package cmd
+package backendcmd
 
 import (
 	"context"
 	"fmt"
 
+	this "github.com/ava-labs/avalanche-cli/pkg/app"
 	"github.com/ava-labs/avalanche-cli/pkg/binutils"
 	"github.com/spf13/cobra"
 )
+
+var app *this.Avalanche
 
 // backendCmd is the command to run the backend gRPC process
 var backendCmd = &cobra.Command{
@@ -17,6 +20,13 @@ var backendCmd = &cobra.Command{
 	Long:  "This tool requires a backend process to run; this command starts it",
 	RunE:  backendController,
 	Args:  cobra.ExactArgs(1),
+}
+
+func SetupBackendCmd(injectedApp *this.Avalanche) *cobra.Command {
+	app = injectedApp
+
+	backendCmd.Hidden = true
+	return backendCmd
 }
 
 func backendController(cmd *cobra.Command, args []string) error {
