+++ cmd/flags/flags.go
@@ -1,0 +1,5 @@
+package flags
+
+const (
+	ForceFlag = "force"
+)
