+++ cmd/subnetcmd/constants.go
@@ -1,0 +1,6 @@
+package subnetcmd
+
+const (
+	subnetEvm = "SubnetEVM"
+	customVM  = "Custom"
+)
