+++ pkg/vm/fees.go
@@ -2,7 +2,7 @@ package vm
 
 import (
 	"github.com/ava-labs/avalanche-cli/cmd/prompts"
-	"github.com/ava-labs/avalanche-cli/ux"
+	"github.com/ava-labs/avalanche-cli/pkg/ux"
 	"github.com/ava-labs/subnet-evm/params"
 )
